refactor(image-cache): split imageCache.ts into focused submodules
imageCache.ts keeps the orchestration entry points + re-exports; the
internals move into imageCache/ with an acyclic dependency graph:

- constants.ts          DB / cache size knobs
- blobCache.ts          in-memory LRU blob map + inflight reads
- urlPool.ts            refcounted shared object URLs
- netFetchScheduler.ts  priority-ordered network fetch slots
- idbStore.ts           IndexedDB open / read / write / evict
- coverSiblings.ts      cover-size sibling probing + upgrade race

Verbatim function-body moves. The only non-move changes are three
named wrappers (cancelScheduledEvict, clearAllUrlEntries, clearCoverState)
that let clearImageCache reach cross-module state — each is the original
lines wrapped in a function, behaviour identical. All seven importers
import from utils/imageCache unchanged.
